pyproject.toml
[project]
name = "person-informations"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "config>=0.5.1",
    "django>=5.2.8",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-role-filters>=1.1.0",
    "icecream>=2.1.8",
    "python-decouple>=3.8",
    "redis>=7.1.0",
]
